Bouge la page reseau admin dans le module correspondant
